# openapi.yaml
openapi: 3.1.0
info:
  title: OceanLaundry API
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  version: 0.3.0
  description: |
    Layanan laundry antar-jemput yang menghubungkan customer, kurir, dan staff
    outlet dalam satu alur: pemilihan paket, penjemputan, penimbangan &
    penetapan harga, pembayaran, pencucian, dan pengantaran Kembali.

    All monetary amounts are integers in the minor unit of Indonesian Rupiah (IDR).
    Clients MUST ignore unknown response fields.
servers:
  - url: https://oceanlaundry.api/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - oauth2: [packages:read]
paths:
...
  /orders:
    get:
      summary: Get orders
      operationId: getOrders
      security:
        - oauth2: [orders:read]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/ValidationError"

    post:
      summary: Create an order
      operationId: createOrder
      security:
        - oauth2: [orders:write]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /orders/{orderId}:
    get:
      summary: Get an order
      operationId: getOrder
      security:
        - oauth2: [orders:read]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}/pickup:
    post:
      summary: Pick up an order
      operationId: pickupOrder
      security:
        - oauth2: [deliveries:write]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      description: Kurir mengambil laundry dari customer.
      responses:
        "200":
          description: Order picked up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /orders/{orderId}/weigh:
    post:
      summary: Weigh an order
      operationId: weighOrder
      security:
        - oauth2: [orders:fulfil]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      description: Staff menimbang laundry dan menetapkan harga. Transitions the order from picked_up to awaiting_payment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - weightGrams
              properties:
                weightGrams:
                  type: integer
                  minimum: 1

      responses:
        "200":
          description: Order weighed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /orders/{orderId}/wash:
    post:
      summary: Start washing an order
      operationId: startWashing
      security:
        - oauth2: [orders:fulfil]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      description: Staff mulai mencuci laundry. 
      responses:
        "200":
          description: Washing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /orders/{orderId}/complete:
    post:
      summary: Complete delivery
      operationId: completeDelivery
      security:
        - oauth2: [deliveries:write]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      description: Kurir menyerahkan laundry kepada customer.
      responses:
        "200":
          description: Order delivered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /orders/{orderId}/cancel:
    post:
      summary: Cancel an order
      operationId: cancelOrder
      security:
        - oauth2: [orders:write]
        - clientCredentials: [orders:write]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      description: |
        Cancels an order that has not yet been paid. Typically called by
        an automated job when an order sits in awaiting_payment too long,
        but may also be usable by a customer or staff.
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401": 
          $ref: "#/components/responses/Unauthorized"
        "403": 
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        Access tokens are issued by the OceanLaundry authorisation server.
        Public clients use Authorization Code + PKCE with no client secret.
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/token
          refreshUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/token
          scopes:
            packages:read: Melihat daftar package
            packages:write: Membuat, mengubah, menghapus package
            orders:read: Membaca order yang terlihat oleh pemanggil
            orders:write: Membuat dan membatalkan order milik sendiri
            orders:fulfil: Menimbang, memulai cuci, menandai siap, menandai selesai
            deliveries:write: Mengambil dan mengantarkan order
            payments:read: Membaca data pembayaran
            payments:write: Membuat pembayaran

    clientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8080/realms/ocean-laundry/protocol/openid-connect/token
          scopes:
            orders:read: Membaca order untuk mencari yang kedaluwarsa (automated job)
            orders:write: Membatalkan order yang kedaluwarsa (automated job)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        UUID version 4 in canonical hyphenated form.
        Required on POST /v1/orders and POST /v1/payments.
        Keys are retained for 24 hours.
        If the same key is reused with an identical body, the server returns
        the previously stored response without processing the request again.
        If the same key is reused with a different body, the server returns
        409 Conflict with type
        https://oceanlaundry.api/problems/idempotency-key-reuse.
      schema:
        type: string
        format: uuid
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    OrderStatus:
      type: string
      enum:
        - placed
        - picked_up
        - weighed
        - awaiting_payment
        - washing
        - ready
        - delivering
        - completed
        - cancelled
    Order:
      type: object
      required:
        - id
        - customerId
        - packageId
        - pickupAddress
        - status
      properties:
        id:
          type: string
          example: ord_001
        customerId:
          type: string
          example: cus_001
        courierId:
          type: [string, "null"]
          example: cur_001
        packageId:
          type: string
          example: pkg_001
        pickupAddress:
          type: string
          example: Jl. Kaliurang No. 10
        status:
          allOf:
            - $ref: "#/components/schemas/OrderStatus"
          example: placed
        weightGrams:
          type: [integer, "null"]
          example: 3500
        totalAmount:
          type: [integer, "null"]
          description: Amount in minor unit of IDR.
          example: 28000
        createdAt:
          type: string
          format: date-time
          example: "2026-08-30T10:00:00Z"
        updatedAt:
          type: [string, "null"]
          format: date-time
          example: "2026-08-30T10:15:00Z"
    CreateOrderRequest:
      type: object
      required: [customerId, packageId, pickupAddress]
      properties:
        customerId:
          type: string
          example: cus_001
        packageId:
          type: string
          example: pkg_001
        pickupAddress:
          type: string
          example: Jln. Jembatan Baru UGM 
...
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
  responses:
    Unauthorized:
      description: The token is absent, unverifiable, or expired.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="invalid_token"'
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The token is valid but does not carry the scope this operation requires.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: 'Bearer error="insufficient_scope", scope="orders:fulfil"'
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Request conflicts with the current state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://oceanlaundry.api/problems/conflict
            title: Order state conflict
            status: 409
            detail: The requested transition is not allowed from the current order status.
            instance: /v1/orders/ord_001
    ValidationError:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://oceanlaundry.api/problems/validation-error
            title: Request validation failed
            status: 422
            detail: weightGrams must be greater than 0.
            instance: /v1/orders/ord_001/weigh
    NotFound:
      description: |
        The object does not exist, or it exists and is not accessible to this caller.
        Both conditions are answered identically so identifiers cannot be enumerated.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://oceanlaundry.api/problems/not-found
            title: Resource not found
            status: 404
            detail: The requested resource was not found.
            instance: /v1/orders/ord_001
...